Clean up and refactoring
