+++ src/interfaces/resource-interfaces/component-resource-interfaces/certifyStatement-interface.ts
