+++ src/interfaces/resource-interfaces/resource-interface.ts
@@ -1,0 +1,6 @@
+import { CertifyStatementIF } from './component-resource-interfaces/certifyStatement-interface'
+
+// Interface to define the resource model example
+export interface ResourceIF {
+  certifyStatementResource: CertifyStatementIF
+}
