+++ src/interfaces/store-interfaces/state-interface.ts
@@ -1,10 +1,8 @@
-import { CertifyStatementIF } from '../resource-interfaces/certifyStatement-interface'
 import { CertifyIF } from '../stepper-interfaces/certify-interface'
 
 // State model example
 export interface StateModelIF {
   stateText: string | undefined,
   currentDate: string,
-  certifyStatementResource: CertifyStatementIF
   certifyState: CertifyIF
 }
