--- src/interfaces/utils-interfaces/resource-interface.ts
@@ -1,6 +1,0 @@
-// Interface to define the resource model example
-export interface ResourceExampleIF {
-  id: number,
-  displayName: string,
-  message: string
-}
