+++ src/store/state/state-model.ts
@@ -3,7 +3,6 @@ import { StateModelIF } from '@/interfaces'
 export const stateModel: StateModelIF = {
   stateText: 'Base State Text',
   currentDate: null,
-  certifyStatementResource: null,
   certifyState: {
     certifyFormValid: false,
     certifiedBy: null
